fix(cli): push hands the plane an ARTIFACT, and the version gate stops lying
`palbase push` could not deploy to a v2 project at all. The palbase-provider arm
uploaded a SOURCE tarball to a deploy ingress that would build it and then polled
a deployment id — measured live 2026-08-24, that ingress answers `not_found`.
This plane's push route is a synchronous relay to the tenant's own management
surface, so the client bundles (where the bundler lives is decided by where the
source enters the system) and the answer IS the outcome: no id, nothing to poll.
The upload/poll machinery goes with it.

The SDK version gate was telling people something no longer true. It said the
cloud runs 18.x and that npm's `latest` points at the 17 line, so its remedy was
"install @palbase/backend@18". Verified 2026-08-24: `latest` is 21.0.0 and the
live tenant runtime resolves 21.0.0 — following that advice would now hand
somebody an SDK three majors behind the runtime. 18 stays as the FLOOR, which is
what it always was; the remedy is `latest`.

And `deploys` printed a 64-character digest into an 8-character column, pushing
every field after it off the line.

Live: centauri pushes 47 endpoints to 8bbwb2pbm through `palbase push`, and
`palbase deploys` shows the row the plane recorded for it.

+++ internal/backend/backend.go
@@ -346,7 +346,11 @@ func deployVersion(v *string) string {
 	if v == nil || *v == "" {
 		return "-"
 	}
-	return *v
+	// SHORTENED, because on this plane a version is an artifact DIGEST — 64 hex
+	// characters, which overruns the column and pushes every field after it off
+	// the line. Twelve is what every other surface prints and what `git log`
+	// taught everyone to read.
+	return short(*v)
 }
 
 // deployWhen renders the createdAt (RFC3339 from the JSON API) as a local
